Fix import of client
Since the Client is located in the following package, I wonder if the import shouldn't be like in this PR.

See also
https://github.com/rabbitmq/hop/blob/master/src/main/java/com/rabbitmq/http/client/Client.java#L17

At least I was not able to import it like stated in the usage guide. However I'm also happy to learn, why it should work like that. Thanks

Cheers