Use -1 for longs in QueueInfo
This commit sets -1 as the default for some long fields in QueueInfo.
This should help make the distinction between an empty queue (e.g. 0
message) and an unavailable queue (e.g. -1 message).

Right now, the value for those fields is 0 when the corresponding JSON
field is not present in the response, which can be confusing.

References #224

Fixes #229