Add HTTP layer isolated tests
To make sure the Client can be used with only required dependencies.

Because of the way Spock/Groovy load classes, HTTP Components is always
required as Client depends on its SSLConnectionSocketFactory class. The
tests could be written with JUnit, but it does not seem possible to run
Spock and JUnit tests with Gradle, so this may have to wait the
migration to Maven.