Bump Spock to 2.0-M4-groovy-3.0
final variables from Groovy tests had to be turned into simple
"def"-declared variables (allowing them seems to be a bug in Groovy
2.4).

The number of run tests is exagerated in the reports: "parameterized"
tests are counted as well as their "parent" method. E.g. for a test with
4 test methods, and each of them with a parameter that has 2 values,
reports count 4x2 + 4 = 12 tests (instead of only 4x2 = 8).