Make http-client dependency mandatory
This is due to the fact that Spring encoder class is not able to
encode special characters such as +, $ or parenthesis needed when we
need to use a regular expression in query parameters (e.g. for paging)
The alternartive was to use URIBuilder from http-client library. But
this implies that it is no longer an optional library but mandatory.