Add support for details objects querying in queue requests
This commit introduces the DetailsParameters class to
query _details objects in the API. For now only queue-related
queries are supported.

The DetailsParameters class allows to set the msg_rates_age and
msg_rates_incr parameters (for messages rates) and the lengths_age
and lengths_incr parameters (for queue lengths). When set up,
information on how count fields have changed is provided (rates and
samples).

Fixes #118